Turret class created 11/30
